move collect_metrics
