+++ benchmarks/benchmark_sampler.py
@@ -11,44 +11,9 @@ from vllm.v1.sample.logits_processor import LogitsProcessors
 from vllm.v1.sample.metadata import SamplingMetadata
 
 from vllm_rbln.v1.sample import WARM_UP_CONFIGS, RBLNSampler
-from vllm_rbln.v1.worker.metrics import PerformanceTracker
+from vllm_rbln.v1.worker.metrics import PerformanceTracker, collect_metrics
 
 MAX_NUM_PROMPT_TOKENS = 64
-
-
-def _collect_metrics(
-    performance_tracker: PerformanceTracker,
-    is_prefill: bool,
-    start_time: float,
-    end_time: float,
-    reports: list[dict],
-    token_count: int,
-) -> None:
-    execution_time = end_time - start_time
-    host_time = None
-    device_time = None
-    ccl_time = None
-    if reports is not None and len(reports) > 0:
-        host_time = reports[0].get("total_host", None)
-        device_time = reports[0].get("total_device", None)
-        ccl_time = reports[0].get("total_ccl", None)
-    if is_prefill:
-        performance_tracker.record_prefill(
-            execution_time,
-            token_count,
-            host_time=host_time,
-            device_time=device_time,
-            ccl_time=ccl_time,
-        )
-    else:
-        performance_tracker.record_decode(
-            execution_time,
-            token_count,
-            host_time=host_time,
-            device_time=device_time,
-            ccl_time=ccl_time,
-        )
-
 
 def _create_penalty_tensor(
     batch_size: int, penalty_value: float, device: torch.device
@@ -236,7 +201,7 @@ def run_benchmark(
         with capture_ctx as model_reports:
             sampler(logits, sampling_metadata)
         # Collect metrics from the sampler
-        _collect_metrics(
+        collect_metrics(
             sampler_performance_tracker,
             is_prefill=False,
             start_time=start_time,
