+++ vllm_rbln/v1/worker/metrics.py
@@ -262,3 +262,36 @@ class PerformanceTracker:
         # Padded decode stats
         self.padded_decode_metrics.show_stats("PADDED DECODE")
         logger.info("=" * 80)
+
+def collect_metrics(
+    performance_tracker: PerformanceTracker,
+    is_prefill: bool,
+    start_time: float,
+    end_time: float,
+    reports: list[dict],
+    token_count: int,
+) -> None:
+    execution_time = end_time - start_time
+    host_time = None
+    device_time = None
+    ccl_time = None
+    if reports is not None and len(reports) > 0:
+        host_time = reports[0].get("total_host", None)
+        device_time = reports[0].get("total_device", None)
+        ccl_time = reports[0].get("total_ccl", None)
+    if is_prefill:
+        performance_tracker.record_prefill(
+            execution_time,
+            token_count,
+            host_time=host_time,
+            device_time=device_time,
+            ccl_time=ccl_time,
+        )
+    else:
+        performance_tracker.record_decode(
+            execution_time,
+            token_count,
+            host_time=host_time,
+            device_time=device_time,
+            ccl_time=ccl_time,
+        )
