+++ vllm_rbln/v1/worker/rbln_model_runner.py
@@ -141,7 +141,7 @@ from vllm_rbln.v1.sample.rbln_rejection_sampler import RBLNRejectionSampler
 from vllm_rbln.v1.spec_decode.eagle import RBLNEagleProposer
 from vllm_rbln.v1.spec_decode.medusa import RBLNMedusaProposer
 from vllm_rbln.v1.worker.bucketing import get_bucketing_manager
-from vllm_rbln.v1.worker.metrics import PerformanceTracker
+from vllm_rbln.v1.worker.metrics import PerformanceTracker, collect_metrics
 
 if TYPE_CHECKING:
     from vllm.model_executor.model_loader.tensorizer import TensorizerConfig
@@ -1860,7 +1860,7 @@ class RBLNModelRunner(LoRAModelRunnerMixin, KVConnectorModelRunnerMixin):
                     sampler_output.sampled_token_ids
                 )
         if envs.VLLM_RBLN_METRICS and self.sampler_performance_tracker is not None:
-            self.collect_metrics(
+            collect_metrics(
                 self.sampler_performance_tracker,
                 self.is_prefills()[0],
                 start_time=sampler_start_time,
@@ -3187,7 +3187,7 @@ class RBLNModelRunner(LoRAModelRunnerMixin, KVConnectorModelRunnerMixin):
 
         if self.e2e_performance_tracker is not None:
             self.e2e_end_time = time.perf_counter()
-            self.collect_metrics(
+            collect_metrics(
                 self.e2e_performance_tracker,
                 is_prefills[0],
                 start_time=self.e2e_start_time,
@@ -4543,41 +4543,6 @@ class RBLNModelRunner(LoRAModelRunnerMixin, KVConnectorModelRunnerMixin):
         return (
             self.speculative_config is not None and self.speculative_config.use_eagle()
         )
-
-    def collect_metrics(
-        self,
-        performance_tracker: PerformanceTracker,
-        is_prefill: bool,
-        start_time: float,
-        end_time: float,
-        reports: list[dict],
-        token_count: int,
-    ) -> None:
-        execution_time = end_time - start_time
-        host_time = None
-        device_time = None
-        ccl_time = None
-        if reports is not None and len(reports) > 0:
-            host_time = reports[0].get("total_host", None)
-            device_time = reports[0].get("total_device", None)
-            ccl_time = reports[0].get("total_ccl", None)
-        if is_prefill:
-            performance_tracker.record_prefill(
-                execution_time,
-                token_count,
-                host_time=host_time,
-                device_time=device_time,
-                ccl_time=ccl_time,
-            )
-        else:
-            performance_tracker.record_decode(
-                execution_time,
-                token_count,
-                host_time=host_time,
-                device_time=device_time,
-                ccl_time=ccl_time,
-            )
-
 
 def create_lora_mask(
     input_ids: torch.Tensor,
